Move Parser interface into yajco-runtime and add ParserFactory
