+++ yajco-modules/yajco-antlr4-parser-generator-module/src/main/resources/yajco/generator/parsergen/antlr4/templates/ParseException.java.vm
@@ -1,6 +1,6 @@
 package $parserPackageName;
 
-public class ParseException extends yajco.generator.parsergen.ParseException {
+public class ParseException extends yajco.parser.ParseException {
     public ParseException(String message, Throwable cause) {
         super(message, cause);
     }
