+++ yajco-modules/yajco-antlr4-parser-generator-module/src/main/resources/yajco/generator/parsergen/antlr4/templates/Parser.java.vm
@@ -6,7 +6,7 @@ import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.BailErrorStrategy;
 import org.antlr.v4.runtime.misc.ParseCancellationException;
 
-public class $parserClassName implements yajco.generator.parsergen.Parser<$mainElementClassName, ParseException> {
+public class $parserClassName implements yajco.parser.Parser<$mainElementClassName, ParseException> {
     @Override
     public $mainElementClassName parse(String input) throws ParseException {
         TokenSource lexer = new ${ANTLRLexerFullClassName}(CharStreams.fromString(input));
@@ -30,6 +30,11 @@ public class $parserClassName implements yajco.generator.parsergen.Parser<$mainE
         } catch (java.io.IOException e) {
             throw new ParseException("Failed to read input file", e);
         }
+    }
+
+    @Override
+    public Class<$mainElementClassName> mainNodeType() {
+        return ${mainElementClassName}.class;
     }
 
     private String readAsString(java.io.Reader r) throws java.io.IOException {
