+++ yajco-modules/yajco-javacc-parser-generator-module/src/main/resources/yajco/generator/parsergen/javacc/templates/ParserException.javavm
@@ -1,6 +1,6 @@
 package $parserPackageName;
 
-public class ParseException extends yajco.generator.parsergen.ParseException {
+public class ParseException extends yajco.parser.ParseException {
   public ParseException(String message, Throwable cause) {
     super(message, cause);
   }
