+++ yajco-parser/src/main/java/yajco/parser/LALRParseException.java
@@ -1,6 +1,6 @@
 package yajco.parser;
 
-public class LALRParseException extends yajco.generator.parsergen.ParseException {
+public class LALRParseException extends ParseException {
     public LALRParseException(String message, Throwable cause) {
         super(message, cause);
     }
