+++ yajco-prototype-modules/yajco-lisa-parser-generator-module/src/main/resources/yajco/generator/parsergen/lisa/templates/ParseExceptionClassTemplate.vm
@@ -1,7 +1,7 @@
 ##Input - parserClassPackageName
 package $parserClassPackageName;
 
-public class LisaParseException extends yajco.generator.parsergen.ParseException {
+public class LisaParseException extends yajco.parser.ParseException {
     public LisaParseException(String message, Throwable cause) {
         super(message, cause);
     }
