+++ yajco-runtime/src/main/java/yajco/parser/ParseException.java
@@ -1,4 +1,4 @@
-package yajco.generator.parsergen;
+package yajco.parser;
 
 public class ParseException extends Exception {
     public ParseException(String message, Throwable cause) {
